+++ core/src/com/unciv/ui/screens/pickerscreens/TileNotePopup.kt
@@ -10,6 +10,7 @@ import com.unciv.ui.components.tilegroups.TileGroup
 import com.unciv.ui.components.tilegroups.TileSetStrings
 import com.unciv.ui.popups.AskTextPopup
 import com.unciv.ui.screens.worldscreen.WorldScreen
+import com.unciv.view.CivView
 import com.unciv.view.TileView
 
 /** Popup for editing a tile note (map pin) */
@@ -24,7 +25,7 @@ fun TileNotePopup(
     AskTextPopup(
         screen = worldScreen,
         label = "Add note for tile".tr(),
-        icon = tile.getTileGroupIcon(),
+        icon = tile.getTileGroupIcon(viewer = worldScreen.getGameViewConsideringForOfWar().civView),
         defaultText = existingNote ?: "",
         maxLength = 64,
         actionOnOk = { note ->
@@ -41,15 +42,15 @@ fun TileNotePopup(
 }
 
 /** Renders the tile's texture set (terrain, resources, improvements, rivers) as a preview icon */
-internal fun Tile.getTileGroupIcon(size: Float = 80f): Group {
+internal fun Tile.getTileGroupIcon(size: Float = 80f, viewer: CivView? = null): Group {
     setTerrainTransients()
     return TileGroup(
-        TileView.forSingleTile(this),
+        TileView.forSingleTile(this, viewer?.getCiv()),
         TileSetStrings(ruleset, UncivGame.Current.settings),
         size * 36f / 54f  // TileGroup normally spills out of its bounding box
     ).apply {
         isForceVisible = true
         isForMapEditorIcon = true
-        update()
+        update(viewer)
     }
 }